Split shell job log wait tests
